requirements.txt
BeautifulSoup==3.2.1
boto==2.38.0
Django==1.8.7
django-bakery==0.7.5
latimes-calculate==0.3.1
rauth==0.7.2
requests==2.8.1
six==1.10.0
wheel==0.24.0
